refactor: extract socket/polling logic into socket/index.js
